no need to clear the formats. We are going to be pasting over them later.
2000-10-15  Jody Goldberg <jgoldberg@home.com>

	* src/clipboard.c (clipboard_paste_region) : no need to clear the
	  formats.  We are going to be pasting over them later.  This means
	  that we don't rerender everything in the region then replace it.

	* src/rendered-value.c (rendered_value_new_ext) : Pass colwidth as a
	  float to get beter resolution.
	* src/format.c (format_value) : Ditto.

	* src/format.c (fmt_general_float) : More tuning.

+++ src/rendered-value.c
@@ -70,7 +70,7 @@ rendered_value_new_ext (Cell *cell, MStyle *mstyle)
 	RenderedValue	*res;
 	Sheet		*sheet;
 	StyleColor	*color;
-	int		 col_width = 0;
+	float		 col_width = 0.;
 	char *str;
 
 	g_return_val_if_fail (cell != NULL, NULL);
@@ -102,9 +102,8 @@ rendered_value_new_ext (Cell *cell, MStyle *mstyle)
 
 				/* FIXME : how does one get the width of the
 				 * widest character used to display numbers.
-				 * Use 4 as the max width for now, and
-				 * intentionally round DOWN to measue the max
-				 * full characters that will fit */
+				 * Use 4 as the max width for now.
+				 */
 				double const font_width = gdk_string_measure
 					(gdk_font, "4");
 				if (font_width > 0)
